requirements.txt
anthropic>=0.34.2
pytrends>=4.9.2
praw>=7.7.1
pymongo>=4.8.0
httpx>=0.27.0
APScheduler>=3.10.4
python-dotenv>=1.0.1
moviepy>=1.0.3
Pillow>=10.4.0
pykokoro>=0.2.1
soundfile>=0.12.1
numpy>=1.26.4
tenacity>=8.3.0
